+++ packages/shared/src/schemas/eventLog.schema.ts
@@ -1,0 +1,45 @@
+import {z} from 'zod';
+
+import {Environment, EventType} from '@shared/types/eventLog.types';
+import {FeedItemActionType, FeedItemIdSchema} from '@shared/types/feedItems.types';
+import {FirestoreTimestampSchema} from '@shared/types/firebase.types';
+import {UserFeedSubscriptionIdSchema} from '@shared/types/userFeedSubscriptions.types';
+
+import {AccountIdSchema} from '@shared/schemas/accounts.schema';
+import {ActorSchema} from '@shared/schemas/actors.schema';
+
+/** Zod schema for an {@link EventId}. */
+// TODO: Consider adding `brand()` and defining `EventId` based on this schema.
+export const EventIdSchema = z.string().uuid();
+
+export const FeedItemActionEventLogItemDataSchema = z.object({
+  feedItemId: FeedItemIdSchema,
+  feedItemActionType: z.nativeEnum(FeedItemActionType),
+});
+
+export const FeedItemImportedEventLogItemDataSchema = z.object({
+  feedItemId: FeedItemIdSchema,
+});
+
+export const UserFeedSubscriptionEventLogItemDataSchema = z.object({
+  userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
+});
+
+const EventLogItemDataSchema = FeedItemActionEventLogItemDataSchema.or(
+  UserFeedSubscriptionEventLogItemDataSchema
+).or(FeedItemImportedEventLogItemDataSchema);
+
+/** Zod schema for an {@link EventLogItem} persisted to Firestore. */
+export const EventLogItemFromStorageSchema = z.object({
+  eventId: EventIdSchema,
+  eventType: z.nativeEnum(EventType),
+  accountId: AccountIdSchema,
+  actor: ActorSchema,
+  environment: z.nativeEnum(Environment),
+  data: EventLogItemDataSchema,
+  createdTime: FirestoreTimestampSchema.or(z.date()),
+  lastUpdatedTime: FirestoreTimestampSchema.or(z.date()),
+});
+
+/** Type for an {@link EventLogItem} persisted to Firestore. */
+export type EventLogItemFromStorage = z.infer<typeof EventLogItemFromStorageSchema>;
