+++ packages/shared/src/types/eventLog.types.ts
@@ -4,15 +4,9 @@ import {makeUuid} from '@shared/lib/utils.shared';
 
 import type {AccountId} from '@shared/types/accounts.types';
 import type {Actor} from '@shared/types/actors.types';
-import {FeedItemActionType, FeedItemIdSchema} from '@shared/types/feedItems.types';
-import type {FeedItemId} from '@shared/types/feedItems.types';
-import {FirestoreTimestampSchema} from '@shared/types/firebase.types';
-import {UserFeedSubscriptionIdSchema} from '@shared/types/userFeedSubscriptions.types';
+import type {FeedItemActionType, FeedItemId} from '@shared/types/feedItems.types';
 import type {UserFeedSubscriptionId} from '@shared/types/userFeedSubscriptions.types';
 import type {BaseStoreItem} from '@shared/types/utils.types';
-
-import {AccountIdSchema} from '@shared/schemas/accounts.schema';
-import {ActorSchema} from '@shared/schemas/actors.schema';
 
 /**
  * Strongly-typed type for an event's unique identifier. Prefer this over plain strings.
@@ -49,11 +43,6 @@ export interface FeedItemActionEventLogItemData extends Record<string, unknown> 
   readonly feedItemActionType: FeedItemActionType;
 }
 
-export const FeedItemActionEventLogItemDataSchema = z.object({
-  feedItemId: FeedItemIdSchema,
-  feedItemActionType: z.nativeEnum(FeedItemActionType),
-});
-
 export interface UserFeedSubscriptionEventLogItemData extends Record<string, unknown> {
   readonly userFeedSubscriptionId: UserFeedSubscriptionId;
   // TODO: Add `userFeedSubscriptionActionType`.
@@ -62,18 +51,6 @@ export interface UserFeedSubscriptionEventLogItemData extends Record<string, unk
 export interface FeedItemImportedEventLogItemData extends Record<string, unknown> {
   readonly feedItemId: FeedItemId;
 }
-
-export const FeedItemImportedEventLogItemDataSchema = z.object({
-  feedItemId: FeedItemIdSchema,
-});
-
-export const UserFeedSubscriptionEventLogItemDataSchema = z.object({
-  userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
-});
-
-const EventLogItemDataSchema = FeedItemActionEventLogItemDataSchema.or(
-  UserFeedSubscriptionEventLogItemDataSchema
-).or(FeedItemImportedEventLogItemDataSchema);
 
 /**
  * Base interface for all event log items. Most things that happen in the app are logged and tracked
@@ -91,25 +68,6 @@ interface BaseEventLogItem extends BaseStoreItem {
   /** Arbitrary data associated with the event. */
   readonly data?: Record<string, unknown>;
 }
-
-/**
- * Zod schema for an {@link EventLogItem} persisted to Firestore.
- */
-export const EventLogItemFromStorageSchema = z.object({
-  eventId: EventIdSchema,
-  eventType: z.nativeEnum(EventType),
-  accountId: AccountIdSchema,
-  actor: ActorSchema,
-  environment: z.nativeEnum(Environment),
-  data: EventLogItemDataSchema,
-  createdTime: FirestoreTimestampSchema.or(z.date()),
-  lastUpdatedTime: FirestoreTimestampSchema.or(z.date()),
-});
-
-/**
- * Type for an {@link EventLogItem} persisted to Firestore.
- */
-export type EventLogItemFromStorage = z.infer<typeof EventLogItemFromStorageSchema>;
 
 export interface FeedItemActionEventLogItem extends BaseEventLogItem {
   readonly eventType: EventType.FeedItemAction;
