+++ packages/sharedServer/src/services/eventLog.server.ts
@@ -11,9 +11,11 @@ import {parseEventId, parseEventLogItem} from '@shared/parsers/eventLog.parser';
 
 import type {AccountId} from '@shared/types/accounts.types';
 import {Environment, EventType, makeEventId} from '@shared/types/eventLog.types';
-import type {EventId, EventLogItem, EventLogItemFromStorage} from '@shared/types/eventLog.types';
+import type {EventId, EventLogItem} from '@shared/types/eventLog.types';
 import type {FeedItemId} from '@shared/types/feedItems.types';
 import type {AsyncResult} from '@shared/types/results.types';
+
+import type {EventLogItemFromStorage} from '@shared/schemas/eventLog.schema';
 
 import {
   makeFirestoreDataConverter,
